requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
scikit-learn
python-dotenv
pymongo
passlib[bcrypt]
groq
apscheduler
email-validator